CR: Moved AWS classes to own package, create generic CloudInstance class
