+++ monkey/common/cloud/aws/aws_instance.py
@@ -6,6 +6,8 @@ import logging
 
 __author__ = 'itay.mizeretz'
 
+from common.cloud.instance import CloudInstance
+
 AWS_INSTANCE_METADATA_LOCAL_IP_ADDRESS = "169.254.169.254"
 AWS_LATEST_METADATA_URI_PREFIX = 'http://{0}/latest/'.format(AWS_INSTANCE_METADATA_LOCAL_IP_ADDRESS)
 ACCOUNT_ID_KEY = "accountId"
@@ -13,7 +15,7 @@ ACCOUNT_ID_KEY = "accountId"
 logger = logging.getLogger(__name__)
 
 
-class AwsInstance(object):
+class AwsInstance(CloudInstance):
     """
     Class which gives useful information about the current instance you're on.
     """
@@ -57,7 +59,7 @@ class AwsInstance(object):
     def get_region(self):
         return self.region
 
-    def is_aws_instance(self):
+    def is_instance(self):
         return self.instance_id is not None
 
     @staticmethod
