+++ monkey/common/cloud/aws/aws_service_test.py
