+++ monkey/common/cloud/environment_names.py
@@ -1,7 +1,12 @@
+# When adding a new environment to this file, make sure to add it to ALL_ENV_NAMES as well!
+
 UNKNOWN = "Unknown"
 ON_PREMISE = "On Premise"
 AZURE = "Azure"
 AWS = "AWS"
 GCP = "GCP"
+ALIBABA = "Alibaba Cloud"
+IBM = "IBM Cloud"
+DigitalOcean = "Digital Ocean"
 
-ALL_ENV_NAMES = [UNKNOWN, ON_PREMISE, AZURE, AWS, GCP]
+ALL_ENV_NAMES = [UNKNOWN, ON_PREMISE, AZURE, AWS, GCP, ALIBABA, IBM, DigitalOcean]
