+++ monkey/common/cloud/instance.py
@@ -1,0 +1,3 @@
+class CloudInstance(object):
+    def is_instance(self) -> bool:
+        raise NotImplementedError()
